Committing first draft of API
in this commit I have a basic frontend that will show data recorded for a day in the past of the database. When a PH is sent to the API it automatically inputs the current time as the time entry which I figured would be the most efficient way to deal with time entries. Let me know if you guys have any recommendations.